feat(addons): enable windows users to login
