--- addons/security/bin/templates/login.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Windows Login</title>
-    <style>
-      body {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        min-height: 100vh;
-        background-color: #f0f0f0;
-        margin: 0px;
-      }
-
-      .login-container {
-        width: 400px;
-        border: 1px solid #ccc;
-        padding: 20px;
-        border-radius: 5px;
-        background-color: white;
-        box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-      }
-
-      h2 {
-        text-align: center;
-        margin-bottom: 20px;
-      }
-
-      label {
-        display: block;
-        margin-bottom: 5px;
-      }
-
-      input[type="text"],
-      input[type="password"] {
-        width: 100%;
-        padding: 8px;
-        margin-bottom: 10px;
-        border: 1px solid #ccc;
-        border-radius: 3px;
-        box-sizing: border-box;
-      }
-
-      button {
-        width: 100%;
-        padding: 10px;
-        background-color: #007bff;
-        color: white;
-        border: none;
-        border-radius: 3px;
-        cursor: pointer;
-      }
-
-      button:hover {
-        background-color: #0056b3; /
-      }
-
-      .error-message {
-            color: red;
-            margin-top: 10px;
-            margin-bottom: 10px;
-            text-align: center;
-        }
-    </style>
-  </head>
-  <body>
-    <div class="login-container">
-      <h2>Windows Login</h2>
-      <form method="post">
-        <input type="hidden" name="login_challenge" value="{{.LoginChallenge}}">
-
-        <label for="username">Username:</label>
-        <input type="text" id="username" name="username" required />
-
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="password" required />
-
-        <div class="error-message">{{.Error}}</div>
-
-        <button type="submit">Login</button>
-      </form>
-    </div>
-  </body>
-</html>
