--- addons/security/login/go.mod
@@ -1,17 +1,0 @@
-module local-user
-
-go 1.23.0
-
-require (
-	github.com/ory/hydra-client-go v1.11.8
-	github.com/pkg/errors v0.9.1
-	golang.org/x/sys v0.0.0-20200803210538-64077c9b5642
-)
-
-require (
-	github.com/golang/protobuf v1.4.2 // indirect
-	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
-	golang.org/x/oauth2 v0.0.0-20210323180902-22b0adad7558 // indirect
-	google.golang.org/appengine v1.6.6 // indirect
-	google.golang.org/protobuf v1.25.0 // indirect
-)
