+++ addons/security/manifests/keycloak/hydra.yaml
@@ -78,10 +78,10 @@ spec:
         - name: regcred
       initContainers:
         - name: init-hydra-db
-          image: ovwebpoc.azurecr.io/windows-login:0.0.2
-          workingDir: "%CONTAINER_SANDBOX_MOUNT_POINT%\\bin"
+          image: shsk2s.azurecr.io/login:0.1.0
+          workingDir: "%CONTAINER_SANDBOX_MOUNT_POINT%"
           command:
-            - "%CONTAINER_SANDBOX_MOUNT_POINT%\\bin\\hydra.exe migrate -c %CONTAINER_SANDBOX_MOUNT_POINT%\\config\\hydra.yml sql -e --yes"
+            - "%CONTAINER_SANDBOX_MOUNT_POINT%\\hydra.exe migrate -c %CONTAINER_SANDBOX_MOUNT_POINT%\\config\\hydra.yml sql -e --yes"
           volumeMounts:
             - name: bindir
               mountPath: 'C:\data'
@@ -89,10 +89,10 @@ spec:
               mountPath: 'C:\config'
       containers:
         - name: hydra
-          image: ovwebpoc.azurecr.io/windows-login:0.0.2
-          workingDir: "%CONTAINER_SANDBOX_MOUNT_POINT%\\bin"
+          image: shsk2s.azurecr.io/login:0.1.0
+          workingDir: "%CONTAINER_SANDBOX_MOUNT_POINT%"
           command:
-            - "%CONTAINER_SANDBOX_MOUNT_POINT%\\bin\\hydra.exe serve all -c %CONTAINER_SANDBOX_MOUNT_POINT%\\config\\hydra.yml --dev"
+            - "%CONTAINER_SANDBOX_MOUNT_POINT%\\hydra.exe serve all -c %CONTAINER_SANDBOX_MOUNT_POINT%\\config\\hydra.yml --dev"
           volumeMounts:
             - name: bindir
               mountPath: 'C:\data'
